Add files via upload
Sonar dataset is a real time case study of submarines that is taking measurement of various signals coming from the hitting of rocks or mines that is laid the enemy country. So here 60 nearby fields are given and the machine has to learn and detect whether the hitting object is rock or mine.
So the output/label or target  is the last column  either the signal  coming back is detecting the rock or mine. 
http://archive.ics.uci.edu/ml/datasets/connectionist+bench+(sonar,+mines+vs.+rocks)